requirements.txt
# Core PyTorch dependencies
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

# FastAPI and web service dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
requests>=2.31.0
pydantic>=2.0.0

# MAGI-specific dependencies
accelerate==0.32.1
beautifulsoup4==4.13.4
debugpy==1.8.14
diffusers==0.29.2
einops>=0.6.0
ffmpeg-python
flash-attn==2.4.2
flashinfer-python==0.2.0.post2 --extra-index-url https://flashinfer.ai/whl/cu124/torch2.4/
ftfy==6.2.0
gpustat==1.1.1
imageio==2.34.0
imageio[ffmpeg]
matplotlib==3.10.1
numpy==1.26.4
protobuf==5.28.3
rich==14.0.0
sentencepiece==0.2.0
timm==1.0.15
torchdiffeq==0.2.4
transformers==4.42.3
